+++ cmd/envpilot-core/main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/zhangyehao/envpilot/internal/config"
 )
 
-var version = "0.4.0"
+var version = "0.4.1"
 
 func main() {
 	if err := run(os.Args[1:]); err != nil {